+++ NEWS.md
@@ -16,6 +16,7 @@
 ## Bug fixes
 
 * `view_df()` did not work with double values (with decimal points) when `show.values = TRUE`.
+* `view_df()` caused an error when a variable has completely missing values.
 * `plot_models()` did not properly remove intercepts from output for survey models, when `show.intercept = FALSE`.
 * `plot_models()` did not automatically transform axis for all applicable model types.
 * `get_model_data()` did not work for marginal effects plots.